+++ test/settings.test.js
@@ -1,0 +1,51 @@
+const loadDefaultSettings = require('../src/settings/loadDefaultSettings');
+
+describe('Test loadDefaultSettings', () => {
+    it('Test undefined settings', () => {
+        const settings = loadDefaultSettings();
+        expect(settings).toStrictEqual({
+            fatalErrorOnCreate: false,
+            mapIfNotFound: false,
+            ingnoreOnTranslate: [],
+            unlinkInputObject: false,
+            defaultGetResponse: undefined,
+            defaultGetAllResponse: []
+        });
+    });
+    it('Custom settings will be loaded correctly', () => {
+        const settings = loadDefaultSettings({
+            fatalErrorOnCreate: true,
+            mapIfNotFound: true,
+            ingnoreOnTranslate: [null, undefined],
+            unlinkInputObject: true,
+            defaultGetResponse: '',
+            defaultGetAllResponse: null
+        });
+        expect(settings).toStrictEqual({
+            fatalErrorOnCreate: true,
+            mapIfNotFound: true,
+            ingnoreOnTranslate: [null, undefined],
+            unlinkInputObject: true,
+            defaultGetResponse: '',
+            defaultGetAllResponse: null
+        });
+    });
+    it('Custom settings of invalid type will be set to default', () => {
+        const settings = loadDefaultSettings({
+            fatalErrorOnCreate: undefined,
+            mapIfNotFound: null,
+            ingnoreOnTranslate: {},
+            unlinkInputObject: 'true',
+            defaultGetResponse: '',
+            defaultGetAllResponse: null
+        });
+        expect(settings).toStrictEqual({
+            fatalErrorOnCreate: false,
+            mapIfNotFound: false,
+            ingnoreOnTranslate: [],
+            unlinkInputObject: false,
+            defaultGetResponse: '',
+            defaultGetAllResponse: null
+        });
+    });
+});